follow float precision

+++ src/main/java/page/anim/AnimBox.java
@@ -34,9 +34,9 @@ interface AnimBox {
 
 	void update();
 
-	void setSiz(double siz);
+	void setSiz(float siz);
 
-	double getSiz();
+	float getSiz();
 
 	void draw();
 }
@@ -50,7 +50,7 @@ class BufferedAnim extends Canvas implements AnimBox {
 
 	private EAnimD<?> ent;
 
-	private double siz = 0.5;
+	private float siz = 0.5f;
 
 	protected BufferedAnim() {
 		setIgnoreRepaint(true);
@@ -127,12 +127,12 @@ class BufferedAnim extends Canvas implements AnimBox {
 	}
 
 	@Override
-	public void setSiz(double siz) {
+	public void setSiz(float siz) {
 		this.siz = siz;
 	}
 
 	@Override
-	public double getSiz() {
+	public float getSiz() {
 		return siz;
 	}
 }
@@ -144,7 +144,7 @@ class GLAnimBox extends GLCstd implements AnimBox {
 	int[] c = new int[] { 70, 140, 160 };
 	int[] f = new int[] { 85, 185, 205 };
 
-	private double siz = 0.5;
+	private float siz = 0.5f;
 
 	private EAnimD<?> ent;
 
@@ -214,12 +214,12 @@ class GLAnimBox extends GLCstd implements AnimBox {
 	}
 
 	@Override
-	public void setSiz(double siz) {
+	public void setSiz(float siz) {
 		this.siz = siz;
 	}
 
 	@Override
-	public double getSiz() {
+	public float getSiz() {
 		return siz;
 	}
